[Fix] concernsのディレクトリ構成を変更
